Complete Level-2: Add, List
